requirements.txt
barcode==1.0.4
chardet==5.2.0
Levenshtein==0.25.0
pillow==10.2.0
python-Levenshtein==0.25.0
rapidfuzz==3.6.1
reportlab==4.1.0
